// clients/admin-ui/src/types/api/models/CustomTaxonomyColor.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export enum CustomTaxonomyColor {
  TAXONOMY_WHITE = 'taxonomy_white',
  TAXONOMY_RED = 'taxonomy_red',
  TAXONOMY_ORANGE = 'taxonomy_orange',
  TAXONOMY_YELLOW = 'taxonomy_yellow',
  TAXONOMY_GREEN = 'taxonomy_green',
  TAXONOMY_BLUE = 'taxonomy_blue',
  TAXONOMY_PURPLE = 'taxonomy_purple',
  SANDSTONE = 'sandstone',
  MINOS = 'minos',
}